Day3 Code
Piece Select Panel showing and pieces are centered in their controls.
Mouse Over on the board gets the where the player would like to play a
piece. Mouse click events are added to piece select panel, but have not
been changed to get the piece. When next turn is called, players rotate
around board, changing the color and the pieces in the piece select
panel.